Update checkov
Running this action yields a deprecation warning from 2022:

> DNS / Plan Cloud Infrastructure
The `set-output` command is deprecated and will be disabled soon. Please upgrade to using Environment Files. For more information see: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/

The issue was reported & resolved in https://github.com/bridgecrewio/checkov-action/issues/105. 

The recommended way to use the action is to specify only the major version (12) so we pick up fixes as they're released, so this change removes the build version pinning. This will resolve the deprecation issue by pulling in a new version (2 years newer)